requirement.txt
numpy
scipy
matplotlib
py3gpp
ipykernel
paho-mqtt

